# packages/contracts/openapi.v1.yaml
openapi: 3.1.0
info:
  title: The Penthouse API
  version: 1.0.0-alpha
  description: Private social platform API. All protected endpoints require a Bearer access token.
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          oneOf:
            - type: string
              description: Human-readable error message
            - type: object
              description: Zod field-level validation error (fieldErrors + formErrors)
    RegisterRequest:
      type: object
      required: [username, password, inviteCode]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
          pattern: '^[a-z0-9._-]+$'
          description: Trimmed and lowercased by the server
        password:
          type: string
          minLength: 10
          maxLength: 128
          description: Must not start or end with spaces
        inviteCode:
          type: string
          minLength: 6
          maxLength: 64
          description: Trimmed and uppercased by the server
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
          pattern: '^[a-z0-9._-]+$'
          description: Trimmed and lowercased by the server
        password:
          type: string
          minLength: 10
          maxLength: 128
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          minLength: 20
          description: Opaque refresh token (96-char hex from server)
    PasswordResetRequest:
      type: object
      required: [username, recoveryCode, newPassword]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
          pattern: '^[a-z0-9._-]+$'
          description: Trimmed and lowercased by the server
        recoveryCode:
          type: string
          minLength: 16
          maxLength: 24
          description: Signup recovery code. Hyphens and spaces are ignored.
        newPassword:
          type: string
          minLength: 10
          maxLength: 128
          description: Must not start or end with spaces
    UserRole:
      type: string
      enum: [admin, member]
...
    AuthUser:
      type: object
      required: [id, username, displayName, avatarUrl, role, mustChangePassword]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
          nullable: true
        role:
          $ref: '#/components/schemas/UserRole'
        mustChangePassword:
          type: boolean
    AuthResponse:
      type: object
      required: [user, accessToken, refreshToken]
      properties:
        user:
          $ref: '#/components/schemas/AuthUser'
        accessToken:
          type: string
          description: Short-lived JWT (default 15m)
        refreshToken:
          type: string
          description: Long-lived opaque token for rotation (default 7d)
        recoveryCode:
          type: string
          description: One-time recovery code display for signup, reset, or backfilled older accounts
...
paths:
...
  /api/v1/auth/register:
    post:
      summary: Register with invite code
      operationId: authRegister
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registered and authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error or invalid/expired invite code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidInvite:
                  value:
                    error: 'Invalid or expired invite code'
                validationError:
                  value:
                    error:
                      fieldErrors:
                        username: ['String must contain at least 3 character(s)']
                      formErrors: []
        '409':
          description: Username already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: 'Username already exists'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/login:
    post:
      summary: Login
      operationId: authLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: 'Invalid credentials'
  /api/v1/auth/refresh:
    post:
      summary: Rotate refresh token and issue new access token
      operationId: authRefresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New token pair issued, old refresh token invalidated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: 'Invalid or expired refresh token'
  /api/v1/auth/password-reset:
    post:
      summary: Reset password using username and recovery code
      operationId: authPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '200':
          description: Password reset complete and authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid username or recovery code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: 'Invalid recovery code or username'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/logout:
    post:
      summary: Logout current refresh token
      operationId: authLogout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '204':
          description: Logged out (refresh token deleted)
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...